rename components and enable main menu links
