+++ sites/frontend/components/Header/Nav/MainMenu/MainMenuColumns/index.js
@@ -3,16 +3,15 @@ import { styled } from '@guardian/guui';
 
 import { desktop, leftCol, wide } from '@guardian/pasteup/breakpoints';
 
-import MainMenuPillar from './MainMenuPillar';
-import pillars from '../../Pillars/__config__';
+import MainMenuColumn from './MainMenuColumn';
+import { columnsConfig } from '../../../Nav/__config__';
 
-const MainMenuPillars = styled('ul')({
+const MainMenuColumns = styled('ul')({
     boxSizing: 'border-box',
     maxWidth: 'none',
     [desktop]: {
         maxWidth: 980,
         backgroundColor: '#e9eff1',
-        marginTop: -20,
         padding: '0 20px',
         position: 'relative',
         margin: '0 auto',
@@ -25,12 +24,12 @@ const MainMenuPillars = styled('ul')({
         maxWidth: 1300,
     },
 });
-MainMenuPillars.displayName = 'MainMenuPillars';
+MainMenuColumns.displayName = 'MainMenuColumns';
 
 export default () => (
-    <MainMenuPillars role="menubar" tabindex="-1">
-        {pillars.map(pillar => (
-            <MainMenuPillar pillar={pillar} key={pillar.label} />
+    <MainMenuColumns role="menubar" tabindex="-1">
+        {columnsConfig.map(column => (
+            <MainMenuColumn column={column} key={column.id} />
         ))}
-    </MainMenuPillars>
+    </MainMenuColumns>
 );
